+++ rentplace/src/main/java/kattsyn/dev/rentplace/services/VerificationCodeService.java
@@ -1,6 +1,6 @@
 package kattsyn.dev.rentplace.services;
 
-import kattsyn.dev.rentplace.dtos.CodeResponse;
+import kattsyn.dev.rentplace.dtos.responses.CodeResponse;
 
 public interface VerificationCodeService {
     boolean validateCode(String email, String code);
